# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentDialog Agent API",
    "version": "dev",
    "description": "The surface an AI agent integrates against: conversations, messages, human queries and webhooks."
  },
  "servers": [
    {
      "url": "https://api.agentdialog.io"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/agent/conversations": {
      "post": {
        "tags": [
          "conversations"
        ],
        "summary": "Start a conversation",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Repeat a POST with the same key and the original response comes back instead of the work happening twice. Only successful responses are remembered, so a refusal leaves the key free for the corrected retry.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "maxLength": 256
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2048
                  },
                  "context": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "intentType": {
                    "type": "string",
                    "enum": [
                      "permission",
                      "clarification",
                      "solicitation",
                      "notification"
                    ]
                  },
                  "settings": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The conversation, created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "createdByAgentId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "title": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "description": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "active",
                            "archived",
                            "closed"
                          ]
                        },
                        "context": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "intentType": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "enum": [
                            "permission",
                            "clarification",
                            "solicitation",
                            "notification",
                            null
                          ]
                        },
                        "settings": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "createdByAgentId",
                        "title",
                        "description",
                        "status",
                        "context",
                        "intentType",
                        "settings",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/IdempotencyKeyReused"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "description": "The request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "tags": [
          "conversations"
        ],
        "summary": "List your conversations",
        "description": "Paginated, but not with the API's usual envelope: this route builds its own pagination object with only hasMore and count, and never returns nextCursor or prevCursor even though the cursor query param is accepted.",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversations the authenticated agent created, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "createdByAgentId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "title": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "description": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "archived",
                              "closed"
                            ]
                          },
                          "context": {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          "intentType": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "enum": [
                              "permission",
                              "clarification",
                              "solicitation",
                              "notification",
                              null
                            ]
                          },
                          "settings": {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "createdByAgentId",
                          "title",
                          "description",
                          "status",
                          "context",
                          "intentType",
                          "settings",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "hasMore": {
                          "type": "boolean"
                        },
                        "count": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "hasMore",
                        "count"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "description": "The `limit` or `cursor` query parameter failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An agent API key, prefixed `mge_ag_`."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "NOT_FOUND",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "CONFLICT",
                  "VALIDATION_ERROR",
                  "RATE_LIMIT",
                  "IDEMPOTENCY_IN_PROGRESS",
                  "IDEMPOTENCY_KEY_REUSED",
                  "UNDECIDABLE_QUERY",
                  "PAYLOAD_TOO_LARGE",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              },
              "retryAfter": {
                "type": "integer"
              },
              "reason": {
                "type": "string"
              },
              "detail": {
                "type": "string"
              },
              "remedy": {
                "type": "string"
              },
              "prior_query_id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "The bearer token is missing, malformed, or does not match a live agent.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many requests. `retryAfter` in the error envelope is the number of seconds to wait before trying again.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "IdempotencyKeyReused": {
        "description": "The Idempotency-Key header was reused, either by a request still in flight or by one whose original request had a different body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The request body exceeded the server's configured maximum size (see app.ts's bodyLimit).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "InternalError": {
        "description": "An unexpected error. The error envelope carries no further detail than `message`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    }
  }
}